kernel.spec: remove /lib/modules/%{KernelVer} if empty when uninstall
When uninstall kernel package, we meet this error message:
"ls: cannot access '/lib/modules/5.10.0-xxx": No such file or directory"

check directory before list directory contents.

+++ kernel.spec
@@ -674,7 +674,9 @@ if [ -x %{_sbindir}/weak-modules ]
 then
     %{_sbindir}/weak-modules --remove-kernel %{KernelVer} || exit $?
 fi
-if [ "`ls -A  /lib/modules/%{KernelVer}`" = "" ]; then
+
+# remove empty directory
+if [ -d /lib/modules/%{KernelVer} ] && [ "`ls -A  /lib/modules/%{KernelVer}`" = "" ]; then
     rm -rf /lib/modules/%{KernelVer}
 fi
